Moved PlayCustom, added an interface and fixed some unreliable code in CustomGameMainPanel
