+++ src/main/java/use_cases/custom_game/custom_game_file_management/IPLayCustom.java
@@ -1,0 +1,9 @@
+package use_cases.custom_game.custom_game_file_management;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public interface IPLayCustom extends ActionListener {
+    @Override
+    void actionPerformed (ActionEvent e);
+}
